add unit tests for build tx with op_return opcode
